Move `SelectionLayer` into `TouchScreen`
