+++ src/components/screen/BaseScreen.jsx
@@ -6,7 +6,7 @@ import {
   selectReaderContents,
   selectReaderCurrent,
   selectReaderSetting,
-  selectReaderCalculationsTotal, selectReaderContentFormat, selectReaderIsReadyToRead,
+  selectReaderCalculationsTotal, selectReaderContentFormat, selectReaderIsReadyToRead, selectReaderSelection,
 } from '../../redux/selector';
 import PropTypes, { ContentType, CurrentType, SettingType } from '../prop-types';
 import DOMEventConstants from '../../constants/DOMEventConstants';
@@ -16,12 +16,71 @@ import TouchableScreen from './TouchableScreen';
 import { addEventListener, removeEventListener } from '../../util/EventHandler';
 import { getStyledTouchable } from '../styled';
 import { ContentFormat } from '../../constants/ContentConstants';
+import { screenHeight, waitThenRun } from '../../util/BrowserWrapper';
+import { ViewType } from '../..';
 
 export default class BaseScreen extends React.Component {
+  static defaultProps = {
+    onTouched: null,
+    onSelectionChanged: null,
+    onAnnotationTouched: null,
+    children: null,
+  };
+
+  static propTypes = {
+    disableCalculation: PropTypes.bool.isRequired,
+    setting: SettingType.isRequired,
+    current: CurrentType.isRequired,
+    contents: PropTypes.arrayOf(ContentType).isRequired,
+    actionUpdateContent: PropTypes.func.isRequired,
+    actionUpdateContentError: PropTypes.func.isRequired,
+    calculationsTotal: PropTypes.number.isRequired,
+    contentFormat: PropTypes.oneOf(ContentFormat.toList()).isRequired,
+    isReadyToRead: PropTypes.bool.isRequired,
+    onTouched: PropTypes.func,
+    onSelectionChanged: PropTypes.func,
+    onAnnotationTouched: PropTypes.func,
+    children: PropTypes.node,
+    selectable: PropTypes.bool.isRequired,
+    annotationable: PropTypes.bool.isRequired,
+    annotations: PropTypes.array,
+    selection: PropTypes.object,
+  };
+
+  static getDerivedStateFromProps(props) {
+    // todo temporary code: Force to set annotation recalculation time
+    return BaseScreen.recalculateAnnotations(props.annotations, props.setting.viewType);
+  }
+
+  static needAnnotationRender(viewType, selection) {
+    // todo temporary code: Force to set annotation recalculation time
+    if (viewType === ViewType.SCROLL) {
+      const { offset } = Connector.current.getCurrent();
+      const [top, height] = [offset, screenHeight()];
+      const contentIndexesInScreen = Connector.calculations.getContentIndexesInOffsetRange(top - (height * 2), top + height + (height * 2));
+      return contentIndexesInScreen.includes(selection.contentIndex);
+    }
+    const { contentIndex } = Connector.current.getCurrent();
+    return contentIndex === selection.contentIndex;
+  }
+
+  static recalculateAnnotations(annotations, viewType) {
+    // todo temporary code: Force to set annotation recalculation time
+    return {
+      annotations: annotations.filter(BaseScreen.needAnnotationRender.bind(this, viewType))
+        .map(item => ({
+          ...item,
+          ...Connector.calculations.getAnnotationCalculation(item),
+        })),
+    };
+  }
+
   constructor(props) {
     super(props);
     this.wrapper = React.createRef();
-
+    this.state = {
+      annotations: [],
+    };
     this.onContentLoaded = this.onContentLoaded.bind(this);
     this.onContentError = this.onContentError.bind(this);
     this.onTouchableScreenTouched = this.onTouchableScreenTouched.bind(this);
@@ -87,7 +146,13 @@ export default class BaseScreen extends React.Component {
     }
   }
 
-  moveToOffset() {}
+  moveToOffset() {
+    const { annotations } = this.props;
+    // todo temporary code: Force to set annotation recalculation time
+    waitThenRun(() => {
+      this.setState(BaseScreen.recalculateAnnotations(annotations));
+    }, 0);
+  }
 
   renderContents() { return null; }
 
@@ -100,6 +165,11 @@ export default class BaseScreen extends React.Component {
       contentFormat,
       isReadyToRead,
       children,
+      annotationable,
+      selectable,
+      selection,
+      onSelectionChanged,
+      onAnnotationTouched,
     } = this.props;
 
     return (
@@ -110,6 +180,12 @@ export default class BaseScreen extends React.Component {
         viewType={setting.viewType}
         StyledTouchable={getStyledTouchable(contentFormat, setting.viewType)}
         isReadyToRead={isReadyToRead}
+        annotationable={annotationable}
+        selectable={selectable}
+        annotations={this.state.annotations}
+        selection={selection}
+        onSelectionChanged={onSelectionChanged}
+        onAnnotationTouched={onAnnotationTouched}
       >
         { this.renderContents() }
         { this.renderFooter() }
@@ -119,29 +195,6 @@ export default class BaseScreen extends React.Component {
   }
 }
 
-BaseScreen.defaultProps = {
-  onTouched: null,
-  onSelectionChanged: null,
-  onAnnotationTouched: null,
-  children: null,
-};
-
-BaseScreen.propTypes = {
-  disableCalculation: PropTypes.bool.isRequired,
-  setting: SettingType.isRequired,
-  current: CurrentType.isRequired,
-  contents: PropTypes.arrayOf(ContentType).isRequired,
-  actionUpdateContent: PropTypes.func.isRequired,
-  actionUpdateContentError: PropTypes.func.isRequired,
-  calculationsTotal: PropTypes.number.isRequired,
-  contentFormat: PropTypes.oneOf(ContentFormat.toList()).isRequired,
-  isReadyToRead: PropTypes.bool.isRequired,
-  onTouched: PropTypes.func,
-  onSelectionChanged: PropTypes.func,
-  onAnnotationTouched: PropTypes.func,
-  children: PropTypes.node,
-};
-
 export const mapStateToProps = state => ({
   setting: selectReaderSetting(state),
   current: selectReaderCurrent(state),
@@ -149,6 +202,7 @@ export const mapStateToProps = state => ({
   calculationsTotal: selectReaderCalculationsTotal(state),
   contentFormat: selectReaderContentFormat(state),
   isReadyToRead: selectReaderIsReadyToRead(state),
+  selection: selectReaderSelection(state),
 });
 
 export const mapDispatchToProps = dispatch => ({
