+++ src/constants/StyledConstants.js
@@ -6,8 +6,6 @@ const CHAPTER_INDICATOR_ID_PREFIX = 'ridi_link_c';
 const CHAPTER_ID_PREFIX = 'ridi_c';
 const DEFAULT_FONT = 'os_specific';
 
-const SELECTION_LAYER_EXPANDED_WIDTH = 10;
-
 export {
   DEFAULT_MAX_WIDTH,
   DEFAULT_VERTICAL_MARGIN,
@@ -16,5 +14,4 @@ export {
   CHAPTER_INDICATOR_ID_PREFIX,
   CHAPTER_ID_PREFIX,
   DEFAULT_FONT,
-  SELECTION_LAYER_EXPANDED_WIDTH,
 };
