+++ src/redux/reducer.js
@@ -88,7 +88,6 @@ const invalidateCalculations = state => new ImmutableObjectBuilder(state)
   .set(path.isReadyToRead(), false)
   .set(path.contentsCalculations(), state.calculations.contents.map(s => initialContentCalculationsState(s.index)))
   .set(path.footerCalculations(), initialFooterCalculationsState())
-  .set(path.annotationCalculations(), {})
   .build();
 
 const updateCalculationsTotal = (state, action) => new ImmutableObjectBuilder(state)
@@ -102,10 +101,6 @@ const setReadyToRead = (state, { isReadyToRead }) => new ImmutableObjectBuilder(
 
 const updateSelection = (state, { selection }) => new ImmutableObjectBuilder(state)
   .set(path.selection(), selection)
-  .build();
-
-const updateAnnotationCalculation = (state, { annotationId, calculation }) => new ImmutableObjectBuilder(state)
-  .set(path.annotationCalculation(annotationId), calculation)
   .build();
 
 export default ({
@@ -128,6 +123,5 @@ export default ({
     [actions.UPDATE_FOOTER_CALCULATIONS]: updateFooterCalculation,
     [actions.UPDATE_CALCULATIONS_TOTAL]: updateCalculationsTotal,
     [actions.UPDATE_SELECTION]: updateSelection,
-    [actions.UPDATE_ANNOTATION_CALCULATIONS]: updateAnnotationCalculation,
   });
 };
