+++ src/service/readerjs/ReaderJsHelper.js
@@ -3,6 +3,7 @@ import { isExist } from '../../util/Util';
 import { screenHeight, screenWidth } from '../../util/BrowserWrapper';
 import { EMPTY_READ_LOCATION, READERJS_CONTENT_WRAPPER, ViewType } from '../../constants/SettingConstants';
 import Connector from '../connector';
+import { RectsUtil } from '../../util/SelectionUtil';
 
 const DETECTION_TYPE = 'top'; // bottom or top
 
@@ -35,10 +36,6 @@ class ReaderJsHelper {
 
   get content() {
     return this.readerJs.content;
-  }
-
-  get node() {
-    return this.content.wrapper;
   }
 
   _isValid() {
@@ -92,6 +89,20 @@ class ReaderJsHelper {
 
   getNodeLocationOfCurrentPage() {
     return this.readerJs.getNodeLocationOfCurrentPage(DETECTION_TYPE);
+  }
+
+  getContentRelativeRects(rects) {
+    const { viewType } = Connector.setting.getSetting();
+    const isScroll = viewType === ViewType.SCROLL;
+
+    // todo check result value
+    const result = new RectsUtil(rects.toAbsolute(this.content.wrapper))
+      .translateX(isScroll ? 0 : Connector.setting.getContainerHorizontalMargin() * 2)
+      // .translateY(isScroll ? Connector.setting.getContainerVerticalMargin() + (Connector.setting.getScrollingContentGap() / 2) : 0)
+      .translateY(isScroll ? this.content.wrapper.parentNode.offsetTop + (Connector.setting.getScrollingContentGap() / 2) : 0)
+      .getRects();
+    // console.log(this.content.wrapper, this.content.wrapper.isConnected, isScroll, rects, result);
+    return result;
   }
 }
